+++ job_registry_next.go
@@ -47,40 +47,3 @@ func (c *Client) JobsNext(param *JobsNextParam, header *JobsNextHeader) (*JobsNe
 
 	return &data, nil
 }
-
-type JobsResultParam struct {
-	WorkerID string
-}
-
-type JobsResultRequest struct {
-	JobID string `json:"jobId,omitempty"`
-	Data  struct {
-		Error   bool   `json:"error,omitempty"`
-		Raw     string `json:"raw,omitempty"`
-		Payload any    `json:"payload,omitempty"`
-	} `json:"data,omitempty"`
-}
-
-func (c *Client) JobsResult(param *JobsResultParam, req *JobsResultRequest) error {
-	reqBody, err := sonic.Marshal(req)
-	if err != nil {
-		return err
-	}
-
-	resp, err := c.req.Post(c.apiURL+"/api/jobs-registry/"+param.WorkerID+"/result", "application/json", reqBody)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return ErrorResponse(body)
-	}
-
-	return nil
-}
