+++ worker_join_test.go
@@ -1,0 +1,61 @@
+package oasm
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/go-retryablehttp"
+)
+
+func TestClient_WorkerJoin(t *testing.T) {
+	type fields struct {
+		req    *retryablehttp.Client
+		apiURL string
+		apiKey string
+	}
+	tests := []struct {
+		name   string
+		fields fields
+		// want    *WorkerJoinResponse
+		wantErr bool
+	}{
+		{
+			name: "TestError",
+			fields: fields{
+				req:    retryablehttp.NewClient(),
+				apiURL: "http://localhost:6276",
+				apiKey: "test",
+			},
+			wantErr: true,
+		},
+		{
+			name: "TestSuccess",
+			fields: fields{
+				req:    retryablehttp.NewClient(),
+				apiURL: "https://3pnb3328-5173.asse.devtunnels.ms",
+				apiKey: "LRvageiyjX8boc6OyApx4nigiJSAexXxfzpo",
+			},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			c := &Client{
+				req:    tt.fields.req,
+				apiURL: tt.fields.apiURL,
+				apiKey: tt.fields.apiKey,
+			}
+			got, err := c.WorkerJoin()
+			if (err != nil) != tt.wantErr {
+				t.Errorf("WorkerJoin() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+
+			fmt.Println(got)
+
+			//if tt.want.Scope != got.Scope || tt.want.Type != got.Type {
+			//	t.Errorf("WorkerJoin() got = %v, want %v", got, tt.want)
+			//}
+		})
+	}
+}
